remove putCommencementDate from endpoints
