+++ app/pages/sections/commencement-date/controller.test.js
@@ -1,7 +1,4 @@
-import { putData } from 'buying-catalogue-library';
-import { getCommencementDateContext, putCommencementDate, validateCommencementDateForm } from './controller';
-import { logger } from '../../../logger';
-import { orderApiUrl } from '../../../config';
+import { getCommencementDateContext, validateCommencementDateForm } from './controller';
 import * as contextCreator from './contextCreator';
 import { getCommencementDate } from '../../../helpers/api/ordapi/getCommencementDate';
 
@@ -11,12 +8,6 @@ jest.mock('./contextCreator', () => ({
   getContext: jest.fn(),
 }));
 jest.mock('../../../helpers/api/ordapi/getCommencementDate');
-
-const mockData = {
-  'commencementDate-day': '21',
-  'commencementDate-month': '12',
-  'commencementDate-year': '2020',
-};
 
 describe('commencement-date controller', () => {
   describe('getCommencementDateContext', () => {
@@ -98,88 +89,6 @@ describe('commencement-date controller', () => {
         const errors = validateCommencementDateForm({ data });
 
         expect(errors).toEqual([commencementDateRequired]);
-      });
-    });
-  });
-
-  describe('putCommencementDate', () => {
-    describe('with errors', () => {
-      it('should return error.respose.data if api request is unsuccessful with 400', async () => {
-        const responseData = { errors: [{}] };
-        putData
-          .mockRejectedValueOnce({ response: { status: 400, data: responseData } });
-
-        const response = await putCommencementDate({
-          orderId: 'order-id', data: mockData, accessToken: 'access_token',
-        });
-
-        expect(response).toEqual(responseData);
-      });
-
-      it('should throw an error if api request is unsuccessful with non 400', async () => {
-        putData
-          .mockRejectedValueOnce({ response: { status: 500, data: '500 response data' } });
-
-        try {
-          await putCommencementDate({
-            orderId: 'order-id', data: mockData, accessToken: 'access_token',
-          });
-        } catch (err) {
-          expect(err).toEqual(new Error());
-        }
-      });
-    });
-
-    describe('with no errors', () => {
-      beforeEach(() => {
-        putData.mockReset();
-      });
-
-      it('should call putData once with the correct params', async () => {
-        putData.mockResolvedValueOnce({});
-
-        await putCommencementDate({
-          orderId: 'order-id', data: mockData, accessToken: 'access_token',
-        });
-        expect(putData.mock.calls.length).toEqual(1);
-        expect(putData).toHaveBeenCalledWith({
-          endpoint: `${orderApiUrl}/api/v1/orders/order-id/sections/commencement-date`,
-          body: { commencementDate: '2020-12-21' },
-          accessToken: 'access_token',
-          logger,
-        });
-      });
-
-      it('should call putData once with the correct params when day and month are single digit', async () => {
-        putData
-          .mockResolvedValueOnce({});
-
-        const data = {
-          'commencementDate-day': '1',
-          'commencementDate-month': '2',
-          'commencementDate-year': '2020',
-        };
-
-        await putCommencementDate({
-          orderId: 'order-id', data, accessToken: 'access_token',
-        });
-        expect(putData.mock.calls.length).toEqual(1);
-        expect(putData).toHaveBeenCalledWith({
-          endpoint: `${orderApiUrl}/api/v1/orders/order-id/sections/commencement-date`,
-          body: { commencementDate: '2020-02-01' },
-          accessToken: 'access_token',
-          logger,
-        });
-      });
-
-      it('should return success as true if data is saved successfully', async () => {
-        putData
-          .mockResolvedValueOnce({});
-        const response = await putCommencementDate({
-          orderId: 'order-id', data: mockData, accessToken: 'access_token',
-        });
-        expect(response.success).toEqual(true);
-        expect(response.errors).toEqual(undefined);
       });
     });
   });
